==> scripts/miscellaneous/scripts/06_read_current_score_sheets.Rmd <==
---
title: "read the current scoresheet"
output: html_document
editor_options: 
  chunk_output_type: console
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

```{r}
library(tidyverse)
library(readxl)
library(lubridate)
```

```{r}
# a working code on one example score sheet 
full_path <- 'data/Finished/2021/4-13-2021 SL GVT T cell dose titration.xlsx'
tmp <- excel_sheets(full_path)
example <- tmp %>% 
  set_names(excel_sheets(full_path)) %>% 
  map(~ read_excel(full_path, sheet = .) %>% 
    filter(rowSums(is.na(.)) < ncol(.)) %>% # remove dividing rows that are all NA
    rename_all(~ gsub(" |\\(|\\)|/", "_", .)) %>% 
    rename_all(~ gsub("#$", "num", .)) %>% 
    fill(Group_num) %>% # fill the NA group num
      mutate(Group_num = as.character(Group_num)) %>% 
    select(Group_num:Weight, Posture:Activity,Died_Sac ) %>% 
    mutate(Died_Sac = as_date(Died_Sac))) %>% 
  bind_rows(.id = 'date_scored') %>% 
  mutate(file_name = str_replace_all(full_path, '^.+/','')) %>% 
  mutate(transplant_date = str_extract(file_name, "[0-9]*\\-[0-9]*\\-[0-9]*")) %>% 
  mutate(transplant_date = mdy(transplant_date)) %>% 
  mutate(date_scored = str_replace_all(date_scored, '\\(.+\\)','')) %>% 
  mutate(date_scored = mdy(date_scored))
```
```{r}
input_sheet <- list.files('../data/Finished/', 
                      pattern = '^\\d.+xlsx$', full.names = T, recursive = T)

clean_score_sheet <- function(f_path){
  cleaned = excel_sheets(f_path) %>% 
      set_names(excel_sheets(f_path)) %>% 
      map(~ read_excel(f_path, sheet = .) %>% 
            mutate(across(everything(), as.character)) %>% 
            filter(rowSums(is.na(.)) < ncol(.)) %>% # remove dividing rows that are all NA
            rename_all(~ gsub(" |\\(|\\)|/", "_", .)) %>% 
            rename_all(~ gsub("#$", "num", .)) %>% 
            fill(Group_num) %>% # fill the NA group num
            filter(!str_detect(Group_num, '^Group')) %>% # some has a header line in the middle of it
            filter(!is.na(Mouse_num)) %>% 
            select(Group_num:Weight, Posture:Activity,Died_Sac ) ) %>% 
      bind_rows(.id = 'date_scored') %>% 
      mutate(file_name = str_replace_all(f_path, '^.+/','')) %>% 
      mutate(transplant_date = str_extract(file_name, "[0-9]*\\-[0-9]*\\-[0-9]*")) 
  return(cleaned)
}
# mutate(Died_Sac = as_date(Died_Sac))%>% 
#      mutate(transplant_date = mdy(transplant_date)) %>% 
#      mutate(date_scored = mdy(date_scored))
```


```{r}
# solving the error ones   
clean_finding_error = safely(.f = clean_score_sheet, otherwise = NULL, quiet = T) 

res <- input_sheet %>% 
  set_names(input_sheet) %>% 
  map(~ clean_finding_error(.)) 

error_ones <- map(res, "error") %>% 
  keep(~ !is.null(.))





res_df <- res %>% 
  map(~ pluck(.x = ., 'result')) %>% 
  bind_rows(.id = 'fullpath')
res_df %>% 
  write_csv('../data/res_df.csv')

res_df %>% distinct(fullpath) %>% nrow


```


```{r}
res_df <- read_csv('../data/res_df.csv', col_types = cols(.default = col_character()))

getwd()

res_df_cleaned <- res_df %>% 
  filter(!str_detect(date_scored, '\\(|\\)')) %>% 
  mutate(Group_num = str_replace_all(Group_num, '\\(|\\)',''))

s2021 <- res_df_cleaned %>% 
  filter(str_detect(fullpath, 'Finished//2021'))
```


```{r}
date_scored <- res_df_cleaned %>% 
  count(date_scored)
# there are letters
date_scored_e <- date_scored %>% 
  filter(str_detect(date_scored, '[:alpha]'))
# no letters but the digits are wrong 
date_scored_try <- date_scored %>% 
  filter(!str_detect(date_scored, '[:alpha]')) %>% 
  mutate(datescored = mdy(date_scored))

res_df_cleaned %>%
  filter(date_scored == '7-')
#######################################
Group_num <- res_df_cleaned %>% 
  count(Group_num)

questions <- res_df_cleaned %>% 
  filter(Group_num == '60')



#######################################
Mouse_num <- res_df_cleaned %>% 
  count(Mouse_num)


#######################################
Baseline_Weight <- res_df_cleaned %>% 
  count(Baseline_Weight)

# C009119 : should be ignored 
# N/A and NA some has baseline weight in another col

#######################################
Weight <- res_df_cleaned %>% 
  count(Weight)
# FD, dead, ., NA


#######################################
Posture <- res_df_cleaned %>% 
  count(Posture)

#######################################
Fur <- res_df_cleaned %>% 
  count(Fur)

#######################################
Skin <- res_df_cleaned %>% 
  count(Skin)


#######################################
Activity <- res_df_cleaned %>% 
  count(Activity)

#######################################
Died_Sac <- res_df_cleaned %>% as_tibble() %>%
  mutate(Died_Sac_neworigin =ifelse(grepl("^4\\d{4}", Died_Sac), as.character(as.Date(as.numeric(Died_Sac), origin = "1899-12-30")), Died_Sac)) %>% 
  mutate(date_scored_ymd =  parse_date_time(date_scored, c("ymd", "mdY", "mdy", "BdY", "Bdy", "bdY", "bdy"))) 

Died_Sac_df <- Died_Sac %>% 
  group_by(Died_Sac_neworigin) %>%
   count(Died_Sac)

#  identify those with easy-to-parse date
Died_Sac_tmp <- Died_Sac %>% 
  mutate(
    sac_date_easy = parse_date_time(Died_Sac_neworigin, c("ymd", "mdY", "mdy", "BdY", "Bdy", "bdY", "bdy")),
    year=year(date_scored_ymd))

Died_Sac_done_simple_date <- Died_Sac_tmp %>% 
  filter(!is.na(sac_date_easy)) %>%
  rename(fixed_Died_Sac_date=sac_date_easy)

#  identify those with full dates burried in the column, and mark those with string months

Died_Sac_tmp2 <- Died_Sac_tmp %>% 
  filter(grepl("[0-9]", Died_Sac )) %>% 
  group_by(fullpath, date_scored, year) %>% 
  filter(is.na(sac_date_easy)) %>% 
  mutate(  
    has_str_month = grepl("Jan|Feb|Mar|Apr|May|Ju|Aug|Sep|Oct|Nov|Dec", Died_Sac),
    full_date = gsub("^-", "", gsub("-$", "", gsub("--", "-", gsub("/", "-", gsub("[\\(\\)A-Za-z'\", .:]", "", Died_Sac))))),
    has_full_date = str_count(full_date, "-") == 2) %>% ungroup()

Died_Sac_done_had_full_date <- Died_Sac_tmp2 %>% 
  filter(has_full_date) %>%
  mutate(
    fixed_Died_Sac_date = parse_date_time(full_date, c("ymd", "mdY", "mdy"))
  )


#  use the year from the score date to fix up those with string months

Died_Sac_tmp3 <- Died_Sac_tmp2 %>% 
  filter(!has_full_date) %>%
  mutate(
    str_month = gsub(".*(Jan|Feb|Mar|Apr|May|Jun|Jul|Aug|Sep|Oct|Nov|Dec).*","\\1",  Died_Sac),
    noalpha = gsub(" *$", "", gsub("^ *", "", gsub("> 25%", "", gsub("[A-Za-z'\",();]", "", Died_Sac)))),
    just_day = gsub("[-./()'\" ,]", "", noalpha))

Died_Sac_done_str_date <- Died_Sac_tmp3 %>% 
  filter(has_str_month) %>% 
  mutate(
    raw_date_if_str_mo = paste0(year, "-", str_month, "-", just_day),
        fixed_Died_Sac_date =  parse_date_time(raw_date_if_str_mo, orders=c("%Y-%b-%d"))
  )



Died_Sac_tmp4 <- Died_Sac_tmp3 %>%
  filter(!has_str_month) %>% 
  mutate(fixed_Died_Sac_date =parse_date_time(paste0(year, "/", noalpha), orders=c( "ymd", "mdY", "mdy", "BdY")))

Died_Sac_done_month_day <- Died_Sac_tmp4 %>% 
  filter(!is.na(fixed_Died_Sac_date)) 


#############3  gather them all together
grep("_done_", ls(), value = T)
these_cols <- c("fullpath", "date_scored", "Group_num", "Mouse_num", "file_name", "fixed_Died_Sac_date")

fixed_Died_Sac <- rbind(
  Died_Sac_done_had_full_date %>% select(all_of(these_cols)),
Died_Sac_done_simple_date%>% select(all_of(these_cols)),
Died_Sac_done_month_day%>% select(all_of(these_cols)),
Died_Sac_done_str_date%>% select(all_of(these_cols)))




#######################################
transplant_date <- res_df_cleaned %>% 
  count(transplant_date) %>% 
  mutate(transplant_date_corrected = mdy(transplant_date)) %>% 
  split(is.na(.$transplant_date_corrected)) 


transplant_date_co <- bind_rows(
  transplant_date %>% pluck('FALSE'),
  transplant_date %>% pluck('TRUE') %>% mutate(transplant_date_corrected = ymd(transplant_date))
)
# few problems it seems 

#######################################
Strain <- res_df_cleaned %>% 
  count(Strain)
# do we keep this col?

#######################################
Baseline_Wt <- res_df_cleaned %>% 
  count(`Baseline_Wt.`)

Baseline_Weight_num <- res_df_cleaned %>% 
  filter(is.na(Baseline_Weight)) %>% 
  nrow

Baseline_Wt._num <- res_df_cleaned %>% 
  filter(!is.na(`Baseline_Wt.`)) %>% 
  nrow
# some are in Baseline_Wt column
```


```{r}
# finding the warnings and messages 
clean_finding_warning <- quietly(.f = clean_score_sheet)
warnning_res <- input_sheet %>% 
  pull(fullpath) %>% 
  set_names(input_sheet %>% 
  pull(fullpath)) %>% 
  map(~ clean_finding_warning(.)) 

warnning_res %>%
     map("messages") %>% 
     unlist()


warn <- warnning_res %>%
     map("warnings") %>% 
     unlist() %>% 
  enframe() %>% 
  mutate(name = str_replace(name, 'xlsx.+$','xlsx'))

warn %>% 
  count(name, sort = T)
```

```{r}
full_path <- '../data/Finished/2021//2-19-2021 SJ BM185 CD19 CD20 BCL-2 PD-1 FasBB 1E5.xlsx'
shet <- '7-26-2021'
cleaned = excel_sheets(full_path) %>% 
      set_names(excel_sheets(full_path)) %>% 
      map(~ read_excel(full_path, sheet = .) %>% 
        filter(rowSums(is.na(.)) < ncol(.)) %>% # remove dividing rows that are all NA
        rename_all(~ gsub(" |\\(|\\)|/", "_", .)) %>% 
        rename_all(~ gsub("#$", "num", .)) %>% 
        fill(Group_num) %>% # fill the NA group num
        mutate(Group_num = as.character(Group_num),
               Mouse_num = as.character(Mouse_num)) %>% 
        filter(!str_detect(Group_num, '^Group')) %>% # some has a header line in the middle of it
        filter(!is.na(Mouse_num)) %>% 
        mutate(Baseline_Weight = as.double(Baseline_Weight)) %>% 
        select(Group_num:Weight, Posture:Activity,Died_Sac ) %>%
          #mutate(Died_Sac = )
        mutate(Died_Sac = as_date(Died_Sac))) %>% 
      bind_rows(.id = 'date_scored') %>% 
      mutate(file_name = str_replace_all(full_path, '^.+/','')) %>% 
      mutate(transplant_date = str_extract(file_name, "[0-9]*\\-[0-9]*\\-[0-9]*")) %>% 
      mutate(transplant_date = mdy(transplant_date)) %>% 
      mutate(date_scored = mdy(date_scored))
```
