==> scripts/miscellaneous/scripts/09_duke.Rmd <==
---
title: "Duke samples"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

```{r}
library(tidyverse)
library(vdbR)
connect_database()
get_table_from_database('asv_alpha_diversity_ag')
get_table_from_database('shotgun_lookup_ad')
get_table_from_database('asv_annotation_blast_ag')
dat <- asv_alpha_diversity_ag %>% 
  distinct(oligos_id) %>% 
  mutate(poolid = str_extract(oligos_id, 'pool.+$')) %>% 
  distinct(poolid) %>% 
  filter(str_detect(poolid, 'pool835|pool935|pool940|pool1042|pool1048|pool1050|pool1064|pool1109|pool1117|pool1121'))

#pool1109
asv_alpha_diversity_ag %>% 
  write_csv('../data/asv_alpha_diversity_ag.csv')


```

```{r}
# total: 2678
duke <- read_csv('../data/duke_samples_for_angel_02-22-2022 (1).csv') %>% 
  rename(oligos_id = oligoid)

# already 1769
already <- duke %>% 
  inner_join(asv_alpha_diversity_ag, by = 'oligos_id') %>% 
  distinct(oligos_id, .keep_all = T)

already %>% 
  select(oligos_id, path_pool ) %>% 
  write_csv('../data/duke/already_1769_path.csv')

# missing 
missing <- duke %>% 
  filter(!oligos_id %in% asv_alpha_diversity_ag$oligos_id)

find <- missing %>% 
  filter(submitted_for_sequencing == 'yes') %>% 
  select(sampleid, sampleid_clean, oligos_id, notes , experiment)  %>% 
  mutate(sid = str_replace_all(oligos_id, '-|/|_','\\.'))

find %>% 
  count(notes)

two <- asv_alpha_diversity_ag %>% 
  filter(str_detect(path_pool, 'pool1096\\+1097|pool1100')) %>% 
  #select(oligos_id, path_pool) %>% 
  distinct(oligos_id, .keep_all = T) 
 
length(intersect(two$sid, find$sid))
setdiff(two$sid, find$sid)

two %>% 
  write_csv('../data/duke/two_pools_90_path.csv')


```


```{r}
setdiff( find$sid, two$sid)

# the 228 that don't know what the pool or where the samples are
where <- find %>% 
  filter(!sid %in% two$sid)

where %>% 
  write_csv('../data/duke/where_the_228.csv')
```

```{r}
# reply from John
reply <- read_csv('~/Downloads/where_the_228_JS_03-21-2022.csv') %>% 
  distinct(sid, .keep_all = T) %>% 
  distinct(js_comment)

length(intersect(reply$sid, asv_alpha_diversity_ag$sampleid))

# after running the last pool 11037
reply %>% 
  filter(str_detect(js_comment, '11037'))
```



```{r}
# assemble the file path that need to be copied to the duke folder in amplicon
current <- bind_rows(
  two %>% 
    mutate(R1 = str_glue('{path_pool}/{oligos_id}_R1.fastq.gz'),
           R2 = str_glue('{path_pool}/{oligos_id}_R2.fastq.gz')),
  already %>% 
    mutate(R1 = str_glue('{path_pool}/{oligos_id}_R1.fastq.gz'),
           R2 = str_glue('{path_pool}/{oligos_id}_R2.fastq.gz'))
) %>% 
  filter(!str_detect(oligos_id, '^blank'))


current %>% 
  select(R1, R2) %>% 
  gather() %>% 
  mutate(cmd = str_glue('cp {value} /home/daia1/my_workdir/samples/amplicon/duke')) %>% 
  select(cmd) %>% 
  write_csv('../data/cp_duke_1851.sh',col_names = F)
```

```{r}
# find out the sampleids of those 1851 samples
alpha_all <- asv_alpha_diversity_ag %>% 
  distinct(sampleid, oligos_id, simpson_reciprocal, shannon)

alpha <- alpha_all %>% 
  filter(oligos_id %in% current$oligos_id)

# the oligos ID are correct, the sampleid need some cleaning
clean_sampleid <- alpha %>% 
  mutate(sampleid = if_else(sampleid != oligos_id, str_replace(oligos_id, '\\.\\.pool.+$',''), sampleid))


cts <- get_counts_subset(clean_sampleid$sampleid) %>% 
  select(asv_key:count_total, count_relative)

taxa <- asv_annotation_blast_ag %>% 
  filter(asv_key %in% cts$asv_key) %>% 
  select(asv_key:blast_pass)

list(alpha = alpha,
     counts = cts,
     taxa = taxa) %>% 
  imap(function(.x, .y){
    write_csv(.x, str_glue('../data/Duke_ASV_{.y}.csv'))
  })

```
```{r}
# for the table that match our sampleids to their sampleids:
# first figure out what exactly is their sampleid
see <- already %>% 
  filter(!str_detect(oligos_id, '^blank')) %>% 
  select(oligos_id:timepoint_int) %>% 
  rename(dukeid = sampleid.x)
see %>% 
  write_csv('../data/what_is_duke_sampleid_1761.csv')
```


```{r}
# the ones that need re-sequencing missing909 - two 90 + blank 8 = 827
# the question is the two pools that Emily sequenced the latest
two_cleaned <- two %>% 
  mutate(sampleid_clean = str_replace_all(sampleid, '\\.','_'))
length(intersect(two_cleaned$sampleid_clean, missing$sampleid_clean))
two_cleaned$sampleid_clean
setdiff(two_cleaned$sampleid_clean, missing$sampleid_clean)

# send the sampleid of the "two " to lauren and John for identifying anything
two %>% 
  select(sampleid) %>% 
  write_csv('../data/duke_two_pools_90.csv')
```

```{r}
missing %>% 
  filter(str_detect(notes, 'Emily')) %>% pull(sampleid_clean)
```
