==> scripts/miscellaneous/scripts/10_upload_data.Rmd <==
---
title: "upload data"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

```{r}
library(tidyverse)
```

```{r}
# 16s 
path <- tibble(path = c(
  '/Volumes/vandenBrinkLab/MMF/Sequencing Data/Rena Feinman/Miseq.230/Sample_pool1126_complete',
  '/Volumes/vandenBrinkLab/MMF/Sequencing Data/Rena Feinman/Miseq.230/Sample_pool1127_complete',
  '/Volumes/vandenBrinkLab/MMF/Sequencing Data/Rena Feinman/Miseq.230/Sample_pool1128_complete',
  '/Volumes/vandenBrinkLab/MMF/Sequencing Data/Rena Feinman/Miseq.230/Sample_pool1129_complete',
  
)) 

list.files('/Volumes/vandenbrinklab/MMF/Sequencing Data/Kylee/Project_12892/DIANA_0451')
```
